backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
cloudinary
openai
python-dotenv
